+++ packages/agtree/CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Fixed
+
+- Recognition of `uBO` modifier in a string with several colons.
+
 ## [3.2.1] - 2025-05-28
 
 ### Added